Refactor translations and enhance stat display
